+++ cmd/ssh.go
@@ -39,7 +39,10 @@ func ssh(args []string, config *dao.Config) {
 	core.CheckIfError(err)
 	servers := []dao.Server{*server}
 
-	err = run.ParseServers(&servers)
+	errConnect, err := run.ParseServers(config.SSHConfigFile, &servers)
+	if len(errConnect) > 0 {
+		core.Exit(&errConnect[0])
+	}
 	core.CheckIfError(err)
 
 	err = run.SSHToServer(*server, config.DisableVerifyHost, config.KnownHostsFile)