+++ core/run/exec.go
@@ -1,6 +1,7 @@
 package run
 
 import (
+	"errors"
 	"fmt"
 	"golang.org/x/crypto/ssh"
 	"os"
@@ -11,7 +12,6 @@ import (
 	"sync"
 
 	"github.com/jedib0t/go-pretty/v6/text"
-	"github.com/kevinburke/ssh_config"
 
 	"github.com/alajmo/sake/core"
 	"github.com/alajmo/sake/core/dao"
@@ -60,9 +60,25 @@ func (run *Run) RunTask(
 		return err
 	}
 
-	err = ParseServers(&run.Servers)
+	errConnects, err := ParseServers(run.Config.SSHConfigFile, &run.Servers)
 	if err != nil {
 		return err
+	}
+
+	if len(errConnects) > 0 {
+		parseOutput := dao.TableOutput{
+			Headers: []string{"server", "host", "user", "port", "error"},
+			Rows:    []dao.Row{},
+		}
+
+		for _, u := range errConnects {
+			parseOutput.Rows = append(parseOutput.Rows, dao.Row{Columns: []string{u.Name, u.Host, u.User, strconv.Itoa(int(u.Port)), u.Reason}})
+		}
+
+		options := print.PrintTableOptions{Theme: task.Theme, OmitEmpty: task.Spec.OmitEmpty, Output: task.Spec.Output, SuppressEmptyColumns: false}
+		print.PrintTable("Parse Errors", parseOutput.Rows, options, parseOutput.Headers[0:1], parseOutput.Headers[1:])
+
+		return &core.ExecError{Err: errors.New("Parse Error"), ExitCode: 4}
 	}
 
 	err = run.ParseTask(configEnv, userArgs, runFlags, setRunFlags)
@@ -355,85 +371,136 @@ func (run *Run) CleanupClients() {
 }
 
 // ParseServers resolves host, port, proxyjump in users ssh config
-func ParseServers(servers *[]dao.Server) error {
+func ParseServers(sshConfigFile *string, servers *[]dao.Server) ([]ErrConnect, error) {
+	if sshConfigFile == nil {
+		return nil, nil
+	}
+
+	cfg, err := core.ParseFile(*sshConfigFile)
+	if err != nil {
+		return nil, err
+	}
+
+	var errConnects []ErrConnect
 	for i := range *servers {
+		serv := cfg[(*servers)[i].Host]
+
 		// Bastion resolve:
 		//  1. proxyjump alias
 		//	2. proxyjump
 		//	3. bastion alias
-		if proxyJump := ssh_config.Get((*servers)[i].Host, "ProxyJump"); proxyJump != "" {
-			if hostName := ssh_config.Get(proxyJump, "HostName"); hostName != "" {
+		if proxyJump := serv.ProxyJump; proxyJump != "" {
+			if hostName := cfg[proxyJump].HostName; hostName != "" {
 				// 1. proxyjump alias
-				user := ssh_config.Get(proxyJump, "User")
-				if user != "" {
-					(*servers)[i].BastionUser = user
-				} else {
-					(*servers)[i].BastionUser = (*servers)[i].User
-				}
+				(*servers)[i].BastionHost = hostName
 
-				port := ssh_config.Get(proxyJump, "Port")
+				port := cfg[proxyJump].Port
 				if port != "" {
 					p, err := strconv.ParseInt(port, 10, 16)
 					if err != nil {
-						return err
+						errConnect := &ErrConnect{
+							Name:   (*servers)[i].Name,
+							User:   (*servers)[i].User,
+							Host:   (*servers)[i].Host,
+							Port:   (*servers)[i].Port,
+							Reason: err.Error(),
+						}
+						errConnects = append(errConnects, *errConnect)
+						continue
 					}
 					(*servers)[i].BastionPort = uint16(p)
 				} else {
 					(*servers)[i].BastionPort = (*servers)[i].Port
 				}
 
-				(*servers)[i].BastionHost = hostName
+				user := cfg[proxyJump].User
+				if user != "" {
+					(*servers)[i].BastionUser = user
+				} else {
+					(*servers)[i].BastionUser = (*servers)[i].User
+				}
 			} else {
 				// 2. proxyjump
 				user, host, port, err := core.ParseHostName(proxyJump, (*servers)[i].User, (*servers)[i].Port)
 				if err != nil {
-					return err
+					errConnect := &ErrConnect{
+						Name:   (*servers)[i].Name,
+						User:   (*servers)[i].User,
+						Host:   (*servers)[i].Host,
+						Port:   (*servers)[i].Port,
+						Reason: err.Error(),
+					}
+					errConnects = append(errConnects, *errConnect)
+					continue
 				}
 
-				(*servers)[i].BastionUser = user
-				(*servers)[i].BastionPort = port
 				(*servers)[i].BastionHost = host
-			}
-		} else if bastionHost := ssh_config.Get((*servers)[i].BastionHost, "HostName"); bastionHost != "" {
-			// 3. bastion alias
-
-			user := ssh_config.Get((*servers)[i].BastionHost, "User")
-			if user != "" {
+				(*servers)[i].BastionPort = port
 				(*servers)[i].BastionUser = user
-			} else {
-				(*servers)[i].BastionUser = (*servers)[i].User
 			}
+		} else if bastionHost := cfg[(*servers)[i].BastionHost].HostName; bastionHost != "" {
+			// 3. bastion alias
+			(*servers)[i].BastionHost = bastionHost
 
-			port := ssh_config.Get((*servers)[i].BastionHost, "Port")
+			port := cfg[(*servers)[i].BastionHost].Port
 			if port != "" {
 				p, err := strconv.ParseInt(port, 10, 16)
 				if err != nil {
-					return err
+					errConnect := &ErrConnect{
+						Name:   (*servers)[i].Name,
+						User:   (*servers)[i].User,
+						Host:   (*servers)[i].Host,
+						Port:   (*servers)[i].Port,
+						Reason: err.Error(),
+					}
+					errConnects = append(errConnects, *errConnect)
+					continue
 				}
 				(*servers)[i].BastionPort = uint16(p)
 			} else {
 				(*servers)[i].BastionPort = (*servers)[i].Port
 			}
 
-			(*servers)[i].BastionHost = bastionHost
+			user := cfg[(*servers)[i].BastionHost].User
+			if user != "" {
+				(*servers)[i].BastionUser = user
+			} else {
+				(*servers)[i].BastionUser = (*servers)[i].User
+			}
 		}
 
-		host := ssh_config.Get((*servers)[i].Host, "HostName")
+		// HostName
+		host := serv.HostName
 		if host != "" {
 			(*servers)[i].Host = host
 		}
 
-		port := ssh_config.Get((*servers)[i].Host, "Port")
-		if port != "22" {
+		// User
+		user := serv.User
+		if user != "" {
+			(*servers)[i].User = user
+		}
+
+		// Port
+		port := serv.Port
+		if port != "" {
 			p, err := strconv.ParseInt(port, 10, 16)
 			if err != nil {
-				return err
+				errConnect := &ErrConnect{
+					Name:   (*servers)[i].Name,
+					User:   (*servers)[i].User,
+					Host:   (*servers)[i].Host,
+					Port:   (*servers)[i].Port,
+					Reason: err.Error(),
+				}
+				errConnects = append(errConnects, *errConnect)
+				continue
 			}
 			(*servers)[i].Port = uint16(p)
 		}
 	}
 
-	return nil
+	return errConnects, err
 }
 
 func (run *Run) ParseTask(configEnv []string, userArgs []string, runFlags *core.RunFlags, setRunFlags *core.SetRunFlags) error {